pyproject.toml
[project]
name = "main-agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-groq>=0.3.2",
    "pymupdf>=1.26.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
]
